refactor: simplify php error tracking internals
